requirements.txt
attrs==21.4.0
distlib==0.3.4
dnspython==2.1.0
filelock==3.4.2
futures==3.0.5
grpcio==1.43.0
grpcio-tools==1.43.0
iniconfig==1.1.1
mongoengine==0.23.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.22.0
packaging==21.3
pandas==1.3.5
platformdirs==2.4.1
pluggy==1.0.0
protobuf==3.19.2
py==1.11.0
pymongo==3.12.3
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
toml==0.10.2
tomli==2.0.0
types-futures==3.3.7
types-protobuf==3.19.4
typing_extensions==4.0.1
virtualenv==20.13.0
